Updated GET endpoint to accept request query strings
So far we've seen we can send data in the body of a POST request (but also a PUT, DELETE, and any non-GET methods), through a route path variable, and now we'll see one final way data can be send through requests: query strings.

Though any method can use query strings, it's most commonly associated with GET requests because these requests can't have a body.

We have a way to get all users at the moment, but no way to get a specific user. Let's update our endpoint to add to the search criteria from the query strings.

If we console.log the value of request.query, we'll see anything that gets added to the query strings of the request on the format ?name=value&other_name=other_value where the ?, the = and the & are delimiting special characters.